requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1

# ── Configuration & environment ────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# ── Audio analysis ─────────────────────────────────────────────────────────────
librosa>=0.10.0
openai-whisper>=20230314
pydub>=0.25.0
soundfile>=0.12.0

# ── Audio mashup pipeline ──────────────────────────────────────────────────────
demucs>=4.0.0
pyrubberband>=0.3.0
yt-dlp>=2023.3.4
ffmpeg-python>=0.2.0

# ── Vector database (PINNED — do not upgrade) ──────────────────────────────────
chromadb==0.4.22
sentence-transformers>=2.2.0

# ── LLM providers ─────────────────────────────────────────────────────────────
anthropic>=0.18.0
openai>=1.0.0

# ── ML / GPU ──────────────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.6.0
safetensors>=0.4.0
controlnet-aux>=0.0.7
Pillow>=10.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
opencv-python>=4.8.1

# ── Video assembly ─────────────────────────────────────────────────────────────
moviepy>=1.0.3

# ── CLIP for drift detection ───────────────────────────────────────────────────
open-clip-torch>=2.23.0

# ── BLIP-2 for auto-captioning ────────────────────────────────────────────────
# (optional, install separately if needed)
# lavis>=1.0.0

# ── Task persistence ───────────────────────────────────────────────────────────
# SQLite is stdlib — no extra package needed

# ── Testing ────────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
